+++ yt_dlp/extractor/extractors.py
@@ -67,7 +67,10 @@ from .appletrailers import (
     AppleTrailersSectionIE,
 )
 from .applepodcasts import ApplePodcastsIE
-from .archiveorg import ArchiveOrgIE
+from .archiveorg import (
+    ArchiveOrgIE,
+    YoutubeWebArchiveIE,
+)
 from .arcpublishing import ArcPublishingIE
 from .arkena import ArkenaIE
 from .ard import (
@@ -1595,7 +1598,6 @@ from .wdr import (
     WDRElefantIE,
     WDRMobileIE,
 )
-from .webarchive import YoutubeWebArchiveIE
 from .webcaster import (
     WebcasterIE,
     WebcasterFeedIE,
